Update Image.php
Fix error when trying to check file_exists() on NULL value

+++ src/Image.php
@@ -69,7 +69,7 @@ class Image
     {
         imagedestroy($this->resource);
 
-        if (file_exists($this->tempnam)) {
+        if (!is_null($this->tempnam) && file_exists($this->tempnam)) {
             unlink($this->tempnam);
         }
     }
